+++ lib/bsqlite/CHANGELOG.md
@@ -6,9 +6,21 @@ All notable changes to this project will be documented in this file.
 
 _Placeholder_
 
+## [0.1.1] - 2025-02-12
+
+### Changed
+
+-   Improved error message they now also include the query string.
+
+### Added
+
+-   Add many more examples about the different features op the crate.
+-   Add `Connection::open_memory()` function to open an in-memory database.
+
 ## [0.1.0] - 2025-02-10
 
 _Initial release_
 
-[Unreleased]: https://github.com/bplaat/crates/compare/bsqlite%2Fv0.1.0...HEAD
+[Unreleased]: https://github.com/bplaat/crates/compare/bsqlite%2Fv0.1.1...HEAD
+[0.1.1]: https://github.com/bplaat/crates/releases/tag/bsqlite%2Fv0.1.1
 [0.1.0]: https://github.com/bplaat/crates/releases/tag/bsqlite%2Fv0.1.0